audio engine fix

+++ crates/tc-analysis/src/bpm.rs
@@ -2,9 +2,9 @@
 
 /// BPM detector using onset detection and autocorrelation
 pub struct BpmDetector {
-    pub(crate) sample_rate: f64,
+    pub(crate) sample_rate: f32,
     /// Onset detection function history
-    onset_history: std::collections::VecDeque<f64>,
+    onset_history: std::collections::VecDeque<f32>,
     /// Maximum onset history length
     max_history: usize,
     /// Minimum onset history entries before detection is attempted.
@@ -17,7 +17,7 @@ impl BpmDetector {
     /// Create a new BPM detector for the given sample rate.
     ///
     /// Returns an error if `sample_rate` is not positive.
-    pub fn new(sample_rate: f64) -> Result<Self, super::AnalysisError> {
+    pub fn new(sample_rate: f32) -> Result<Self, super::AnalysisError> {
         if sample_rate <= 0.0 {
             return Err(super::AnalysisError::InvalidSampleRate(sample_rate));
         }
@@ -35,11 +35,11 @@ impl BpmDetector {
     /// This is an energy-based approach (not spectral flux); the
     /// onset function measures the root-mean-square energy of each
     /// hop-sized frame.
-    pub fn feed(&mut self, samples: &[(f64, f64)]) {
+    pub fn feed(&mut self, samples: &[(f32, f32)]) {
         let hop_size = 512;
 
         for chunk in samples.chunks(hop_size) {
-            let energy: f64 = chunk.iter().map(|(l, r)| (l * l + r * r) * 0.5).sum();
+            let energy: f32 = chunk.iter().map(|(l, r)| (l * l + r * r) * 0.5).sum();
             let onset = energy.sqrt();
             self.onset_history.push_back(onset);
             if self.onset_history.len() > self.max_history {
@@ -62,7 +62,7 @@ impl BpmDetector {
             };
         }
 
-        let mean: f64 = self.onset_history.iter().sum::<f64>() / self.onset_history.len() as f64;
+        let mean: f32 = self.onset_history.iter().sum::<f32>() / self.onset_history.len() as f32;
 
         // Autocorrelation with mean subtraction (autocovariance)
         let min_lag = (60.0 / 200.0 * self.sample_rate / 512.0) as usize; // 200 BPM max
@@ -80,8 +80,8 @@ impl BpmDetector {
         }
 
         let mut best_lag = min_lag;
-        let mut best_corr = f64::MIN;
-        let mut second_best_corr = f64::MIN;
+        let mut best_corr = f32::MIN;
+        let mut second_best_corr = f32::MIN;
 
         for lag in min_lag..=effective_max_lag {
             let mut corr = 0.0;
@@ -91,7 +91,7 @@ impl BpmDetector {
                 let y = self.onset_history[i + lag] - mean;
                 corr += x * y;
             }
-            corr /= n as f64;
+            corr /= n as f32;
 
             if corr > best_corr {
                 second_best_corr = best_corr;
@@ -102,7 +102,7 @@ impl BpmDetector {
             }
         }
 
-        let bpm = 60.0 * self.sample_rate / 512.0 / best_lag as f64;
+        let bpm = 60.0 * self.sample_rate / 512.0 / best_lag as f32;
 
         // Normalize BPM to 60-200 range (handle double/half time)
         // Use a loop to handle extreme values (e.g., raw BPM of 402 or 25)
@@ -139,7 +139,7 @@ impl BpmDetector {
 
 #[cfg(test)]
 mod tests {
-    use std::f64::consts::PI;
+    use std::f32::consts::PI;
 
     use super::*;
 
@@ -150,11 +150,11 @@ mod tests {
         let sr = 44100.0;
         let beat_interval = (sr * 0.5) as usize; // 120 BPM
         let total_samples = sr as usize * 6; // 6 seconds of audio
-        let samples: Vec<(f64, f64)> = (0..total_samples)
+        let samples: Vec<(f32, f32)> = (0..total_samples)
             .map(|i| {
                 let beat_phase = i % beat_interval;
                 let v = if beat_phase < 512 {
-                    (2.0 * PI * 80.0 * (beat_phase as f64 / sr)).sin() * 0.8
+                    (2.0 * PI * 80.0 * (beat_phase as f32 / sr)).sin() * 0.8
                 } else {
                     0.0
                 };
@@ -183,7 +183,7 @@ mod tests {
         let mut detector = BpmDetector::new(44100.0).unwrap();
         // Feed silence — the autocorrelation will find some lag but
         // the key test is that the normalization loop handles edge cases.
-        let samples: Vec<(f64, f64)> = (0..441000).map(|_| (0.0, 0.0)).collect();
+        let samples: Vec<(f32, f32)> = (0..441000).map(|_| (0.0, 0.0)).collect();
         detector.feed(&samples);
         let result = detector.detect();
         // BPM should be within 60-200 range due to normalization loop
@@ -203,9 +203,9 @@ mod tests {
         // mean-subtracted autocorrelation (no panic, reasonable output)
         let mut detector = BpmDetector::new(44100.0).unwrap();
         // Feed enough samples to exceed min_history threshold
-        let samples: Vec<(f64, f64)> = (0..441000)
+        let samples: Vec<(f32, f32)> = (0..441000)
             .map(|i| {
-                let t = i as f64 / 44100.0;
+                let t = i as f32 / 44100.0;
                 let v = (2.0 * PI * 440.0 * t).sin() * 0.5;
                 (v, v)
             })
