+++ crates/tc-engine/src/analysis/waveform.rs
@@ -7,7 +7,7 @@ use super::AnalysisError;
 #[derive(Debug, Clone)]
 pub struct WaveformData {
     /// (min, max) pairs for each pixel
-    pub peaks: Vec<(f64, f64)>,
+    pub peaks: Vec<(f32, f32)>,
     pub samples_per_pixel: usize,
     pub total_frames: usize,
 }
@@ -29,20 +29,20 @@ impl WaveformGenerator {
     }
 
     /// Generate waveform data from stereo samples
-    pub fn generate(&self, samples: &[(f64, f64)]) -> WaveformData {
+    pub fn generate(&self, samples: &[(f32, f32)]) -> WaveformData {
         let spp = self.samples_per_pixel;
         let num_peaks = samples.len().div_ceil(spp);
         let mut peaks = Vec::with_capacity(num_peaks);
 
         for chunk in samples.chunks(spp) {
-            let mut min = f64::MAX;
-            let mut max = f64::MIN;
+            let mut min = f32::MAX;
+            let mut max = f32::MIN;
             for (l, r) in chunk {
                 let mono = (l + r) * 0.5;
                 min = min.min(mono);
                 max = max.max(mono);
             }
-            if min == f64::MAX {
+            if min == f32::MAX {
                 min = 0.0;
                 max = 0.0;
             }
